refactor(store): one structural comparator for both slices, not two
#4478 landed `applySlots` with its own `deepEqual`, while `chatSlice` already
carried a byte-identical `jsonEqual` for message `meta`/`variants`. Two private
copies of the same comparator is two sets of guarantees to keep in sync, and the
guarantees are the whole point: key-order independence (one side is a fresh
server payload, the other may be an object an in-place reducer has since patched,
and a patch can append a key the payload spells earlier) and field-agnosticism (a
comparator listing a type's fields stops seeing a newly added one and reports
equal on a payload that actually changed, pinning stale content on screen). If
one copy drifts, the drift is silent and the failure mode is a correctness bug in
whichever slice lost the property.

Extracts the comparator to `utils/structuralEqual.ts` as the single `jsonEqual`
and points both slices at it. No behaviour change: the two implementations were
already semantically identical, so this is a dedup, not a fix.

Records the rule in frontend-conventions.md alongside the merge-don't-replace
section that motivates it, and names the two dashboard collections that still
assign wholesale (`sseSubagentStatus`, `sseStatus`) as known gaps rather than
counterexamples, so the section is not read as a description of the store.
